move jersey filters to .resource package
