+++ src/main/java/uk/gov/register/presentation/app/PresentationApplication.java
@@ -25,10 +25,11 @@ import org.glassfish.hk2.utilities.binding.AbstractBinder;
 import org.glassfish.jersey.server.ServerProperties;
 import org.skife.jdbi.v2.DBI;
 import uk.gov.organisation.client.GovukOrganisationClient;
-import uk.gov.register.presentation.ContentSecurityPolicyFilter;
-import uk.gov.register.presentation.ContentTypeOptionsFilter;
+import uk.gov.register.presentation.resource.CacheNoTransformFilterFactory;
+import uk.gov.register.presentation.resource.ContentSecurityPolicyFilter;
+import uk.gov.register.presentation.resource.ContentTypeOptionsFilter;
 import uk.gov.register.presentation.EntryConverter;
-import uk.gov.register.presentation.XssProtectionFilter;
+import uk.gov.register.presentation.resource.XssProtectionFilter;
 import uk.gov.register.presentation.config.FieldsConfiguration;
 import uk.gov.register.presentation.config.PresentationConfiguration;
 import uk.gov.register.presentation.config.PublicBodiesConfiguration;
