+++ src/main/java/uk/gov/register/presentation/resource/ContentSecurityPolicyFilter.java
@@ -1,4 +1,4 @@
-package uk.gov.register.presentation;
+package uk.gov.register.presentation.resource;
 
 import com.google.common.net.HttpHeaders;
