+++ src/main/java/uk/gov/register/presentation/resource/ContentTypeOptionsFilter.java
@@ -1,4 +1,4 @@
-package uk.gov.register.presentation;
+package uk.gov.register.presentation.resource;
 
 import com.google.common.net.HttpHeaders;
 
@@ -7,9 +7,9 @@ import javax.ws.rs.container.ContainerResponseContext;
 import javax.ws.rs.container.ContainerResponseFilter;
 import java.io.IOException;
 
-public class XssProtectionFilter implements ContainerResponseFilter {
+public class ContentTypeOptionsFilter implements ContainerResponseFilter {
     @Override
     public void filter(ContainerRequestContext requestContext, ContainerResponseContext responseContext) throws IOException {
-        responseContext.getHeaders().add(HttpHeaders.X_XSS_PROTECTION, "1; mode=block");
+        responseContext.getHeaders().add(HttpHeaders.X_CONTENT_TYPE_OPTIONS, "nosniff");
     }
 }
